requirements.txt
streamlit
pymupdf
Pillow
requests